mock-local-controller/requirements.txt
aiohttp
aiokafka
grpcio
grpcio-tools
protobuf